feat: Add dependency review workflow
Add a new workflow file, dependency-review.yaml, to enable dependency review for pull requests. This workflow runs on the latest Ubuntu environment and includes steps to checkout the repository and perform the dependency review using the dependency-review-action. The configuration file for the dependency review is specified as 'DevExpPlatform/required-workflows/policies/dependency-review-config.yml@main'.